Get rid of subtle UB
If calculateBoundingBox is true, me from 11 years ago
thought we didn't need to set {min,max}{X,Y}, since
decodeSimpleGlyph will set them. But I was wrong: passing indeterminate
values to a function is UB, even if they're not used by the function.

At least, this is how clang on macOS (Apple clang version 15.0.0 (clang-1500.3.9.4))
seems to interpret the standard; the present commit therefore fixes a miscompilation
on that platform.